Add update task
Without this patch the rake update task is not implemented and as a
result the build script in the Puppet Win Builder project is not
self-sustaining.

This patch implements the update task such that simply running build.bat
from the Puppet Win Builder archive will pull down the latest HEAD of
the master branch from GitHub.

This makes the build system self-sustaining.