(#12473) Add Install Directory UI
Without this patch there is no UI to specify the installation directory.
This patch adds the UI for both the command line and GUI.  The UI
introduced by this patch is the default and looks like:

![Puppet Install Directory UI](http://links.puppetlabs.com/ftw_installdir.png)

The UI is implemented by linking the WixUIExtension.dll extension into the
installer.  The MSI itself is localized using en-us by default.

The installation directory may be specified on the command line using
something like:

    msiexec /qn /l*v install.txt /i puppet.msi INSTALLDIR="C:\test\puppet"