(#12475) Add all Licenses to MSI EULA Screen
Without this patch the MSI package displayed the default Common Public
License.  This is a problem because this is not our license terms.

This patch adds a license document that contains the following
information:

 * License Summary at the top

Then the full text of the license terms for each component in this
installation:

 * Full text of licenses for all bundled software
 * puppet
 * facter
 * ruby
 * rubyinstaller
 * diff-lcs
 * metaclass
 * mocha
 * rspec
 * rspec-core
 * rspec-expectations
 * rspec-mocks
 * sys-admin
 * win32utils
 * WiX Toolset

The screen looks like:

![Puppet Installer License](http://links.puppetlabs.com/ftw_license.png)