(#12474) Add PUPPET_MASTER_HOSTNAME Remembered Property
Without this patch there are two missing features of the installer.

 1. Choices the user makes are not remembered
 2. The user has no way to specify the Puppet Master hostname

This patch addresses the first issue by employing the "Remember
Property" Pattern [1].  This patch writes the INSTALLDIR and
PUPPET_MASTER_HOSTNAME public properties into the registry.  In the
uninstall, repair, or upgrade scenario these keys are recalled using the
RegistrySearch feature of WiX.

If the user overrides these properties on the command line when
performing a re-installation of the package then the command line
specified value is saved before reading the value from the registry and
restored after reading the value from the registry.  This behavior
allows the recalled values to be explicitly overriden.

There is no GUI for the PUPPET_MASTER_HOSTNAME property.  The only way
to specify this is on the command line as shown in the `rake
windows:install` task.

The registry keys the installer writes look like:

![FTW Registry 1](http://links.puppetlabs.com/ftw_registry1.png)

[1] http://links.puppetlabs.com/ftw_remember_properties