(#12474) Copy dialogs to customize from WiX 3.5 Source
Without this patch it is difficult to customize the WixUI_InstallDir
WixUIExtension the installer is using to present the UI screens when
doing a GUI install.

This patch is a straight copy from the WiX 3.5 source [1] as mentioned
in the article on Customised UI's for WiX at [2].

These files will be modified in subsequent patches.

[1] http://wix.codeplex.com/SourceControl/list/changesets?branch=wix35
[2] http://neilsleightholm.blogspot.com/2008/08/customised-uis-for-wix.html