(#12474) Enable Customization of InstallDirDlg and WelcomeDlg
Without this patch it is not easy to customize the welcome dialog
screen and the installation directory dialog screen.

This patch makes it easy to customize these screens by renaming the
identifiers from the copies of the WiX source files.

The renamed ID's are:

 * InstallDirDlg => PuppetInstallDirDlg
 * WelcomeDlg    => PuppetWelcomeDlg

The package will look and behave the same as before with this patch, but
the source files are now part of this project and easy to customize.

The welcome screen with and without this patch looks like:

![Welcome](http://links.puppetlabs.com/ftw_welcome.png)

The install dir with and without this patch looks like:

![Install Dir](http://links.puppetlabs.com/ftw_installdir.png)