(#12474) Add puppet master hostname and agent certname fields
This patch adds two fields to the "PuppetInstallDirDlg" screen.  The
first ask, "Connect to Puppet Master Hostname:" and defaults to "puppet"
while the second asks, "Certificate Name of this Puppet Agent:" and
defaults to the ComputerName property.  This should be the hostname on
a windows machine.

REVISIT - There is currently a problem where the certificate name field
is not properly set to the default value of the ComputerName property
but instead the literal string "[ComputerName]"  I believe I just need
to change the custom action to set the value rather than the Value
attribute of the Property element.

This patch formalizes the following two properties in the UI and as
user-facing command line options to msiexec.

 * PUPPET_MASTER_HOSTNAME
 * PUPPET_AGENT_CERTNAME

The new install directory looks like:

![Install Dir 1](http://links.puppetlabs.com/ftw_installdirdlg1.png)