(#12474) Set PUPPET_AGENT_CERTNAME to ComputerName conditionally
Without this patch the UI certificate name field was not being set to
the dynamic default value of the ComputerName property.

This patch fixes the problem by adding a custom action that sets the
PUPPET_AGENT_CERTNAME to the value of the ComputerName properly only if
the value has not already be specified on the command line.

This patch results in the Install Directory Dialog looking like this by
default:

![Install Directory Dialog](http://links.puppetlabs.com/ftw_installdirdlg2.png)