(#12521) Add UX Preliminary images
This patch overwrites the stock images from the WiX source with the
images provided by Randall.  These images are for placement only and to
get an idea if 24 bit BMP images are supported by the installer.

These images are not actually used in the installer yet.  A subsequent
patch will update the WiX XML files to uses the images explicitly.