(#12521) Add windows:buildui task to build UI only
Without this patch it takes ~5-10 minutes to build the full 20 MB
puppet.msi package.  This is a PITA when making pixel-level changes to
the placement of controls in the UI.

This patch fixes the problem by adding a new windows:buildui task.  This
task builds a package with only one file - the License.  This has the
same UI as the "real" package but is much faster to build and rapidly
develop against.

To build this special package an environment variable must also be set
like so:

    rake windows:buildui BUILD_UI_ONLY=true

The resulting package will be in `pkg/puppet_ui_only.msi`

This patch also illustrates how to use the WiX Preprocessor to
conditionally include other components.  This will be useful when we
separate out the branding of Puppet FOSS and Puppet Enterprise.