(#12521) Add FPO UX images to MSI Package
Without this patch the resulting MSI package uses the bitmap images
shipped as part of the WixUI library.  This is a problem because we need
the package to use the "for placement only" images provided by Randall
and the UX team.

This patch fixes the problem by setting two WixVariable's;
WixUIBannerBmp and WixUIDialogBmp.

We need only replace the bitmap images in the `wix/ui/Bigmaps/`
directory to update the images used in the installer once this patch is
applied.