(maint) Make the build system use all local images
Without this patch only 2 of the 6 images we can customize are
explicitly referenced in the puppet XML file.  This is confusing because
we've copied all 6 default images from the WiX source.  We overwrote two
of them though.

This patch removes the confusion by explicitly using the local copies of
these images rather than the library images.  The patch has no effect on
the final product but it makes things easier to maintain since
everything is in one place and in version control.