(#12561) Fix batch files located in path with spaces
Without this patch the installed software refuses to run because the
batch files do not properly quote the path to the environment.bat
script.  This is a problem because we want to install to a directory
like %PROGRAMFILES% which has spaces, e.g. `C:\Program Files (x86)`

This patch fixes the problem by quoting the fully qualified path to the
environment setup script.