(#12570) Add run_{puppet,facter}_interactive batch files
Without this patch applied there is no easy way to run Puppet and
Facter from the desktop user interface.  This is a problem because the
product is less usable for someone who is expecting a double click UI
and UX.

This patch helps address the problem by adding two batch files to run
Puppet and Facter interactively.

The batch files use a `PAUSE` statement to prevent the `cmd.exe` window
from flashing on the screen and vanishing before the user has a chance
to review the output.

![Run Facter Interactive](http://links.puppetlabs.com/ftw_msi_facter_interactive_1a.png)