(#12561) Make %PROGRAMFILES% the default INSTALLDIR
It is a problem if we refuse to play nice with the Microsoft ecosystem.
The best practice for applications is to install into a vendor-specific
sub-directory of `%PROGRAMFILES%`.

This patch fixes this problem by making `%PROGRAMFILES%\Puppet
Labs\Puppet` the default installation directory for our package.  This
follows Microsoft best practices and recommendations for application
developers.

This will usually be `C:\Program Files (x86)\Puppet Labs\Puppet` since
Ruby is a 32 bit executable.

The user sees this path reflected in the default destination directory
as per:

![Destination](http://links.puppetlabs.com/ftw_msi_120210_3d.png)