(maint) Add en-us localization from WiX Source (noop)
Without this patch we cannot easily customize the localization strings displayed to
the end user.  This is a problem because we want to tweak the font color
and some of the things that are said.

This patch is preliminary work to make customization of localized
strings easy from within the build system.

This patch has no effect on the product as the localized string is not
yet used by light.  A future patch will need to explicitly configure the
build system to use the localization file using the `light.exe -loc
foo.wxl` flag.