(#11842) Link puppet_en-us.wxl localization into MSI
Without this patch the project specific localization strings we've
copied from the WiX source code are not actually linked into the MSI
packages we build.  This is a problem because we need to customize these
localization strings.

This patch fixes the problem by explicitly linking the localization file
into the MSI being built.