(#11842) Customize Welcome screen as per UX
Without this patch the Welcome dialog is the WiX UI default.  This is a
problem because it does not match the information Randall logged in
(#11842) that says the welcome screen should say:

    The first screen should say:

    This will install Puppet [VERSION] and configure the system to fetch
    configurations [EVERY HOUR].

    This version of Puppet will work with puppet masters running Puppet
    [2.7.4] or higher, including all Puppet Enterprise releases since 2.0.

This patch fixes this problem by customizing the welcome string to match
Randall's guidance.  The screen now looks like what is documented in the
README.