(maint) Automate the process of generating screenshots
Without this patch it's tedious to update the README file with screen
shot images.

This patch automates the process by simply using the `ext/screenshots`
script to look at the 5 most recently updated files in my screenshots
directory.

Unfortunately this is highly customized to my laptop, but it's a good
start.  Trollop should make it easy to generalize the automation.