(maint) Add -arch x86 flag to the candle compiler
Without this patch we're relying on the default behavior of WiX to
produce an architecture specific package.  This isn't really a problem
itself because the default is to produce a 32 bit x86 package, but Josh
pointed out that the modern best practice is to use the -arch flag when
compiling the package objects.

This patch moves the build system more in line with best practices by
explicitly using the -arch x86 flag when compiling the package objects.