(#12576) Create confdir and vardir parent folders
Without this patch the MSI package does not create any folders in
`C:\Documents and Settings\All Users\Application Data`  This is a
problem because Puppet on windows defaults to a confdir of
`%ALLUSERSPROFILE%\Application Data\PuppetLabs\puppet\etc` and a vardir
of `%ALLUSERSPROFILE%\Application Data\PuppetLabs\puppet\var`  If the
parent directory of $vardir does not exist then Puppet cannot create
the directories it needs and fails to run.

This patch fixes the problem by creating empty folders for both confdir
and vardir.