(#12520) Write puppet.conf using IniFile element
Without this patch the MSI installer does not do anything with the
PUPPET_AGENT_CERTNAME and PUPPET_MASTER_HOSTNAME properties other than
write them into the registry.

This is a problem because we need Puppet agent to connect to the master
the user has specified and use the certificate name the user has
specified.

This patch fixes the problem by writing the puppet.conf file using the
IniFile element in the WiX schema [1]

[1] http://wix.sourceforge.net/manual-wix3/wix_xsd_inifile.htm

NOTE this patch currently has a UX issue where the default ComputerName
property value on a Windows machine is upper case.  Puppet refuses to
work with upper case certificate names as per (#1168).

An example of the configuration file we write is:

    [main]
    server=puppet
    certname="wintest"