(maint) Add troubleshooting missing .NET 3.5
Without this patch it might be difficult to troubleshoot an issue where
a Windows machine without .NET 3.5 is being used as the build system.
This is a problem scenario because light and candle exit with no error
message if .NET 3.5 is missing.  This can be difficult to troubleshoot.

This patch fixes the problem by adding a troubleshooting section to the
main README file and specifically calls out this situation.