(#12640) Remove PUPPET_AGENT_CERTNAME from GUI
Without this patch, the GUI portion of the installer presents a field to
enter the agent certificate name.  The field defaults to the
ComputerName which defaults to an all uppercase value.

This is a problem for Puppet because the field fills in the value of the
certname setting in the `puppet.conf` file.  When Puppet runs, the
certname setting cannot contain upper case characters as per #1168.

This patch fixes the problem by removing the question from the GUI
entirely.  The MSI now writes the `certname` setting to `puppet.conf`
only if the end user has specified the PUPPET_AGENT_CERTNAME public
property on the command line.  This fixes the problem because in the
default situation Puppet will use the default setting for the certname
which is the value of the `fqdn` fact.

If the user manually specifies PUPPET_AGENT_CERTNAME then they must do
so as a lower case setting.  If they don't, Puppet will give a useful
error message parsing defaults when it runs for the first time.

Alternative solutions, such as using a Custom Action in the MSI to
downcase the ComputerName are not being employed because the WiX authors
strongly discouraged and do not play nicely with anti virus software.

[1] http://robmensching.com/blog/posts/2007/8/17/Zataoca-Custom-actions-are-generally-an-admission-of-failure