(#12570) Add Start Menu shortcuts for puppet and facter
Without this patch the end user needs to open `cmd.exe`, navigate to the
installation folder, navigate into the bind folder and run the batch
file `run_facter_interactive.bat` or `run_puppet_interactive.bat`

This is a problem because it's annoying to do.

This patch fixes the problem by making it less annoying to run `puppet
agent --test` and `facter --puppet`.  The MSI packages will now create
Start Menu entries to run Facter and Puppet inside cmd.exe.  The batch
files issue a `PAUSE` statement at the end so the cmd.exe windows don't
disappear, giving the user a chance to review the output.

Icons are not provided by the start menu shortcut.

This patch writes a registry key into the HKCU hive.  This is because a
Component in an MSI needs a keypath element, which can be a file or a
registry key.  We don't want to use a file because all of the files we
install are in the "per machine" context.  A start menu item is in the
"per user" context.  If we mix per machine and per user components
inside of the same component group, light will issue critical errors
about doing so.

The use of a component with a registry keypath into HKCU avoids the
problem.

The shortcuts are still written into the "All Users" start menu of
course.