(#12627) Add Puppet Enterprise branding
This patch introduces the ability to produce both Puppet Enterprise and
FOSS branded packages.  Without this patch the build system only
produces a "Puppet" branded package.  This is a problem because we also
want packages branded as Puppet Enterprise.

This change also eliminates the requirement to set environment variables
when building the UI only.  The new build commands are documented in
`README_DEVELOPER.markdown` and now look like:

    rake windows:build
    rake windows:buildenterprise PE_VERSION_STRING=2.5.0

    rake windows:buildui
    rake windows:buildenterpriseui PE_VERSION_STRING=2.5.0.1

This patch also introduces dynamic version number calculations inside
the build system itself.  In the case of Puppet, the version of the
package will automatically be parsed from `git describe`.  In the case
of Puppet Enterprise the version of the package defaults to 2.5.0.0 and
can be overriden using the PE_VERSION_STRING environment variable.

This patch also integrates the package with Microsoft's Add or Remove
Programs feature more clearly.  The ARPPRODUCTICON and ARPHELPLINK
public properties are set to provide an Icon and a clickable URL to get
more information and support.