(#12627) Make it easy to brand bitmaps
Without this patch it would be a bit difficult to customize the bitmap
images used in the graphical installer on a per-brand basis.

This patch makes it easy to customize the bitmaps used by using a
BitmapFolder preprocessor variable set inside the BRANDING condition in
the `puppet.wxi` included file.