(maint) Fix stale objects being used
Without this patch if the PE brand package is built immediately after
the Puppet brand package, the object files are not actually rebuilt.

Since the object files are passed brand specific definitions, we need to
make sure to rebuild the object files themselves so we get the correct
branding.

This patch fixes the problem by always invoking the object file tasks
before invoking the linking task.  The end result is that defined
preprocessor variables are always fresh.