(#12488) Support major upgrades of Puppet and PE
Without this patch the MajorUpgrade feature of Microsoft installer is
not explicitly supported.  This is a problem because we need to let end
users upgrade from Puppet to Puppet Enterprise easily and also upgrade
from one release to another release of the same branding.

This patch explicitly supports the MajorUpgrade feature.  Windows
installer uses the following product attributes to figure out if it is
in an upgrade or install scenario.

 * Id (UUID, set to '*', regenerates every build as it should)
 * Language (Set to 1033 statically, should remain static)
 * UpgradeCode (UUID, MUST REMAIN FIXED, statically defined in puppet.wxi)
 * Version (Calculated by the build system and defined in puppet.wxi)

By default Windows installer does not allow downgrading to an older
version.  We explicitly allow downgrades because Puppet Enterprise
versions will be considered older than actual older versions of Puppet.
For example, consider the following situation:

 * Puppet 2.7.10 installed.
 * Puppet Enterprise 2.5.0 is trying to be installed.

In this situation the installer would refuse to install since 2.5.0 is
considered older than 2.7.10.  We explicitly allow downgrades so the
behavior of this patch is for Windows installer to remove the previous
package before the installation phase.