(maint) Add WiX Util Extension
Without this patch we have no built in facility to manage Shortcut
elements that point to Internet URL's.

We need to manage shortcuts that point to our online documentation.

This patch links the WiX Util library into our project which provides a
number of new elements in the util namespace.  One of these new elements
is the InternetShortcut [1] element.

We may or may not use this additional functionality, but we'll likely
need some aspects of this extension library.  Integration with the
service framework for example.

[1] http://wix.sourceforge.net/manual-wix3/util_xsd_internetshortcut.htm