(#11842) Add start menu shortcuts to online documentation
Without this patch the user does not have a convenient way to get
additional help or information immediately after installing Puppet or
Puppet Enterprise.

It is really difficult for us to change the behavior of the last dialog
screen to include clickable URL's.

This patch fixes the problem by creating a "Documentation" folder in the
Start Menu and adding Shortcut entries inside of this folder.  The
entries have our icons and point at local "*.url" files.

The installer itself uses the IniFile element to create these URL files
that point to our actual documentation.  These files will be updated on
new versions of the package and removed on un-installation.