(#11842) Add Start Command Prompt Shortcut
Without this patch there is no easy way for the user to bring up a
command prompt with Puppet or Puppet Enterprise already setup in the
PATH and RUBYLIB.  This is a problem because if the user would like to
work with the command line they need to manually set PATH and RUBYLIB as
we would.

This patch fixes the problem by creating a shortcut that starts a
command prompt with PATH and RUBYLIB configured correctly for the user.