(maint) Bump minimum version from PE 2.0 to PE 2.5
The welcome screen of the installer lies to the user without this patch.
It states, "This version of Puppet Enterprise will work with puppet
masters running Puppet 2.7 or higher, including all Puppet Enterprise
releases since 2.0"

This is a lie because the version of facts_dot_d.rb included in PE
versions before 2.5 cannot work on Windows.  This is a problem because
there are required custom facts on Windows that cause a PE 2.0 master to
explode.