(#12357) Provide a env_windows_installdir fact
Without this patch it's difficult to know where Puppet has been
installed to on a Windows system.  We need to know this to write Puppet
modules that source the environment.bat script so they can run Puppet or
Facter.

This is a problem for the Scheduled Task resource that needs to
configure `puppet inspect` to run periodically.

This patch fixes the problem by setting an environment variable named
`FACTER_env_windows_installdir` in the `environment.bat` script that
configures the runtime environment for all Puppet executables.