(maint) Remove icons from documentation URL shortcuts
Randall pointed out the Puppet Labs logo icon isn't ideal for the
documentation shortcuts.  This patch fixes the problem by removing the
explicit icons and relying on the default behavior of the OS.

The default behavior for a shortcut pointing to a file with a .URL
extension is to use the icon of the default web browser.  This behavior
will do a better job of communicating the fact these are links to
websites rather than shortcuts to local resources on the system.