(maint) Describe how to update gems to be installed
Without this patch the documentation is a bit out of date because we now
have the rake windows:repack and rake windows:upload tasks.  These tasks
make it much easier and faster to update gems included in the MSI
package.

I'm merging this straight in because it's a documentation only update.