(#9595) Enable autoflush mode
When using a log file destination, puppet does not automatically flush
after each message, which can result in lost messages if the ruby process
terminates unexpectedly. This is more of an issue for windows, because we
don't log to syslog by default.

This commit just enables the autoflush setting in puppet.conf so that
windows agents always autoflush by default.