(#12768) Allow ca_server to be specified on the command line
This commit allows the `ca_server` property to be set on the command line
when installing puppet. For example:

    msiexec /qn /i puppet.msi PUPPET_CA_SERVER=myca.acme.com

If the MSI property is specified, but the ca_server property already
exists in puppet.conf, then it will not be overwritten

If the MSI property is not specified or is blank, then the ca_server
property is not written out.