(#12768) Rename PUPPET_MASTER_{HOSTNAME,SERVER}
The previous commit in this change set for (#12768) adds the MSI
property named PUPPET_CA_SERVER.  This is a problem because it is not
consistent with the previous property for the Puppet Master named
PUPPET_MASTER_HOSTNAME.

PUPPET_CA_SERVER is consistent with the Puppet setting named `ca_server`
however.  This patch fixes the inconsistency by renaming
PUPPET_MASTER_HOSTNAME to PUPPET_MASTER_SERVER which maps to the Puppet
`server` setting.