(#12768) Remember PUPPET_CA_SERVER in the Registry
Without this patch I saw this warning while building the MSI packages:

    warning CNDL1006 : Property 'PUPPET_CA_SERVER' does not contain a
    Value attribute and is not marked as Admin, Secure, or Hidden.  The
    Property element is being ignored.

This patch fixes the warning by reading the default value of the
property from the windows registry.

This patch also updates the MSI to write the value used during initial
installation to the registry.  This might be useful for future upgrade
scenarios.