(#12474) Recall INSTALLDIR on a product specific basis
When installing Puppet Enterprise with Puppet already installed, the
original Puppet specific path is the default.  This means Puppet
Enterprise will likely be installed to the path usually used by Puppet.

This patch fixes the problem by saving and recalling INSTALLDIR on a
product specific basis.

The end result is that the default value for Puppet Enterprise will be
used when upgrading from an existing installation of Puppet.

These paths are:

 * `%ProgramFiles%\Puppet Labs\Puppet`
 * `%ProgramFiles%\Puppet Labs\Puppet Enterprise`