(#12975) Patch the PUPPETVERSION for Puppet Enterprise
Without this patch the puppetversion fact reports the same thing for the
Puppet Enterprise branded installer and the Puppet FOSS installer.

The Puppet Enterprise installer should report a puppetversion similar
to:

    C:\> facter puppetversion
    2.7.12 (Puppet Enterprise 2.5.0)

This patch fixes the problem by directly patching puppet/lib/puppet.rb
to contain the version string we want.

Since we're modifying the staged copy of Puppet, this patch introduces
some order dependency to the build process.  As a result, the
recommended way to build all of the packages it to follow this process:

    rake clean
    rake windows:checkout[refs/tags/2.7.12,refs/tags/1.6.6]
    rake windows:build
    rake windows:buildenterprise

Ensuring Puppet is built before Puppet Enterprise will ensure Puppet
does not receive a patched PUPPETVERSION constant so long as the build
tree is cleaned before anything else happens.