(#12620) Add windows:sign rake task to sign the MSI
Without this patch we have no automated rake tasks to digitally sign the
MSI packages we're building.  This is a problem because we need to
easily sign the packages and each person who builds them shouldn't have
to learn the magic invocation if the `signtool` command.

This patch fixes the problem by adding a new rake task named
`windows:sign`.  This task requires the `puppet.msi` and
`puppetenterprise.msi` packages to be built already and will digitally
sign them using the Puppet Labs code signing cert if it is installed on
the system running the task.

The README_DEVELOPER file is also updated with information about how to
install the code signing certificate on the build machine.