(#13427) Archive files and graph on Windows (PE Only)
Without this patch the default behavior of the archive_files,
archive_file_server and graph puppet.conf settings are used.  This
behavior is false, $server and false respectively.

On other platforms Puppet Enterprise configures archive_files to true,
the archive_file_server to the puppet master hostname, and graph = true.
This is a problem because windows is not consistent with other
platforms.  More importantly, file archiving needs to be turned on for
the diff feature of Compliance to work as expected.

This patch fixes the problem by configuring puppet.conf in a manner
consistent with other Puppet Enterprise platforms.  The Puppet FOSS
package will not configure these options to be consistent with Puppet
FOSS on other platforms.

This patch also manages `graph=true` to help users visualize the
configuration catalog.